CCIPでSetParametersメソッドを呼び出すタスクを追加しました。
また、Monster NFT用のタスクを再整理しました。
