+++ pkgs/backend/outputs/contracts-baseSepolia.json
@@ -3,7 +3,7 @@
     "SampleForwarder": "0x046283C3f76178dEAe802f95D831a4F956130dF4",
     "HelloWorld": "0xB93AE7eaBee6FCEc0ff242deA52528ed69A0a6D5",
     "ETHOnline2024NFT": "0x149920786500a12dA84185df4B4aAABe975Df5f8",
-    "DestinationMinter": "0xA433938c78A781E1f6c79f7F8BE2C200cb6bC046",
-    "MonsterNFT": "0x290C4c22069B6801f2ba587A8cBba87d37d4980C"
+    "DestinationMinter": "0xb27E2b20054b9ce343eA51eEd335891d6D0831ca",
+    "MonsterNFT": "0x02fc35AcD83AB0F7d802224B0A13f1462C2C05EA"
   }
 }
