+++ pkgs/backend/outputs/contracts-chilizTestnet.json
@@ -2,6 +2,7 @@
   "contracts": {
     "SampleForwarder": "0xAa363921A48Eac63F802C57658CdEde768B3DAe1",
     "HelloWorld": "0x2B5914De5D5166eBaa423C92BAb8518c85EAA0cb",
-    "BaseNft": "0xAc0DE957C2474519744270f47DEef57FFF088F60"
+    "BaseNft": "0xAc0DE957C2474519744270f47DEef57FFF088F60",
+    "MonsterNFT": "0xa05Db9C31B6ffB6aB817D346E99095e1c1c8317D"
   }
 }
